scripts/requirements.txt
lightrdf==0.3.1
python-magic==0.4.27
python-slugify==8.0.1
rdflib==6.3.2
requests==2.25.1
rich==13.4.1
